+++ app/src/main/java/woowacourse/shopping/feature/product/recent/RecentProductItemViewHolder.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemRecentBinding
-import woowacourse.shopping.model.RecentProductState
+import woowacourse.shopping.feature.product.recent.model.RecentProductState
 
 class RecentProductItemViewHolder(binding: ViewBinding) : RecyclerView.ViewHolder(binding.root) {
     val binding = binding as ItemRecentBinding
