+++ app/src/main/java/woowacourse/shopping/feature/product/recent/model/RecentProductState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.feature.product.recent.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
