+++ app/src/main/res/layout/item_cart_product.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="cartProduct"
-            type="woowacourse.shopping.model.CartProductState" />
+            type="woowacourse.shopping.feature.cart.model.CartProductState" />
     </data>
 
     <com.google.android.material.card.MaterialCardView
