+++ app/src/main/res/layout/item_product.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="product"
-            type="woowacourse.shopping.model.ProductState" />
+            type="woowacourse.shopping.feature.product.model.ProductState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
